src/main/java/com/study/onlinestore/dao/jdbc/mapper/ProductRowMapper.java
package com.study.onlinestore.dao.jdbc.mapper;

import com.study.onlinestore.entity.Product;

import java.sql.ResultSet;
import java.sql.SQLException;

public class ProductRowMapper {
    public Product mapRow(ResultSet resultSet) throws SQLException {
        Product product = new Product();
        product.setId(resultSet.getInt("id"));
        product.setName(resultSet.getString("name"));
        product.setDescription(resultSet.getString("description"));
        product.setPrice(resultSet.getDouble("price"));
        product.setPicturePath(resultSet.getString("picturePath"));
        return product;
    }
}
